Adicionado Testes unitarios e melhorias diversas no codigo

+++ PlaylistRecommender.Domain/Entities/Weather.cs
@@ -4,11 +4,10 @@ namespace PlaylistRecommender.Domain.Entities
 {
     public class Weather : Entity
     {
-        public Weather(float temperature)
+        public Weather(double temperature)
         {
             this.Temperature = temperature;
         }
-
-        public float Temperature { get; private set; }
+        public double Temperature { get; private set; }
     }
 }
